backend/requirements.txt
fastapi
uvicorn
qdrant-client
openai
python-dotenv
python-multipart
PyPDF2
python-docx
chardet
